# read: inline delta-varint decoding in unmarshalDocumentSections