--- a/api.go
+++ b/api.go
@@ -320,7 +320,7 @@ type Repository struct {
 }
 
 func (r *Repository) UnmarshalJSON(data []byte) error {
-	// We define a new type so that we can use json.Unmarhsal
+	// We define a new type so that we can use json.Unmarshal
 	// without recursing into this same method.
 	type repository *Repository
 	repo := repository(r)