@mpimenov 2022-06-20
> What do you think about a closure which which calls Uvarint and mutates data? This would make the call sites a little clearer in my opinion.

Unfortunately, this slow things down a bit.
Current PR vs closure:

```
name                           old time/op    new time/op    delta
UnmarshalDocSections/10-10       52.7ns ± 1%    88.1ns ± 1%  +67.05%  (p=0.000 n=9+9)
UnmarshalDocSections/100-10       422ns ± 1%     795ns ± 0%  +88.11%  (p=0.000 n=9+9)
UnmarshalDocSections/1000-10     5.28µs ± 1%    9.51µs ± 1%  +80.27%  (p=0.000 n=10+10)
UnmarshalDocSections/10000-10    60.1µs ± 0%    95.3µs ± 2%  +58.41%  (p=0.000 n=10+10)
```

It looks like ```binary.Uvarint``` is getting inlined but the closure calls do not: https://godbolt.org/z/KWMYshrzG